requirements.txt
Flask
numpy
Pillow
tensorflow-cpu
scikit-learn
gunicorn